feat(service): add --next option to update command

+++ packages/service/src/scripts/update.ts
@@ -3,10 +3,11 @@ import * as fs from 'fs';
 import * as https from 'https';
 import { runtimeRoot } from '@vuesion/utils/dist/path';
 import { ensureDirectoryExists } from '@vuesion/utils/dist/fileSystem';
-import { log, logError, logErrorBold, logInfoBold, logSuccess, Result } from '@vuesion/utils/dist/ui';
+import { log, logError, logErrorBold, logInfo, logSuccess, Result } from '@vuesion/utils/dist/ui';
 import { VuesionConfig } from '@vuesion/models';
 
 interface IFile {
+  raw_url: string;
   filename: string;
   status: string;
   previous_filename: string;
@@ -14,22 +15,17 @@ interface IFile {
 
 const vuesionRepo = 'https://api.github.com/repos/vuesion/vuesion';
 const deleteFile = (status: string, filePath: string) => {
-  try {
-    fs.unlinkSync(filePath);
-    logError(`${status}: ${filePath}`);
-  } catch (e) {
-    logErrorBold(e.message);
-  }
+  fs.unlinkSync(filePath);
+  logError(`${status}: ${filePath}`);
 };
 const renameFile = (status: string, oldPath: string, newPath: string) => {
   try {
     fs.renameSync(oldPath, newPath);
-    logInfoBold(`${status}: ${oldPath} --> ${newPath}`);
+    logInfo(`${status}: ${oldPath} --> ${newPath}`);
   } catch (e) {
-    logErrorBold(e.message);
+    logErrorBold(`error: ${oldPath} --> ${newPath}`);
   }
 };
-
 const downloadFile = (status: string, filePath: string, url: string) => {
   ensureDirectoryExists(filePath);
   const file = fs.createWriteStream(filePath);
@@ -60,26 +56,55 @@ const downloadFile = (status: string, filePath: string, url: string) => {
       deleteFile(status, filePath);
     });
 };
+const shouldProcessFile = (diffFile: IFile) => {
+  const foldersToNotSync = [
+    '.circleci/',
+    '.github/',
+    'cypress/',
+    'i18n/',
+    'src/app/config',
+    'src/app/example',
+    'src/app/home',
+    'src/server/routes/CounterRoutes.ts',
+    'src/server/routes/DemoRoutes.ts',
+    'src/static',
+    '.all-contributorsrc',
+    'CHANGELOG.md',
+    'CODE_OF_CONDUCT.md',
+    'LICENSE',
+    'README.md',
+  ];
+  let process = true;
 
-const handleFiles = (diffFiles: IFile[]) => {
+  foldersToNotSync.forEach((folder) => {
+    if (diffFile.filename.toLowerCase().indexOf(folder.toLowerCase()) > -1 && process === true) {
+      process = false;
+    }
+  });
+
+  return process;
+};
+const handleFiles = (diffFiles: IFile[], branch: string) => {
   diffFiles.forEach((diffFile: IFile) => {
-    const dest: string = runtimeRoot(diffFile.filename);
-    const url = `https://raw.githubusercontent.com/vuesion/vuesion/master/${diffFile.filename}`;
+    if (shouldProcessFile(diffFile)) {
+      const dest: string = runtimeRoot(diffFile.filename);
+      const url = `https://raw.githubusercontent.com/vuesion/vuesion/${branch}/${diffFile.filename}`;
 
-    if (diffFile.status === 'removed') {
-      deleteFile(diffFile.status, dest);
-    } else if (diffFile.status === 'renamed') {
-      renameFile(diffFile.status, runtimeRoot(diffFile.previous_filename), dest);
-    } else {
-      downloadFile(diffFile.status, dest, url);
+      if (diffFile.status === 'removed') {
+        deleteFile(diffFile.status, dest);
+      } else if (diffFile.status === 'renamed') {
+        renameFile(diffFile.status, runtimeRoot(diffFile.previous_filename), dest);
+      } else {
+        downloadFile(diffFile.status, dest, url);
+      }
     }
   });
 };
 
-export async function run() {
+export async function run(next = false) {
   try {
-    const tagsResponse: AxiosResponse<any> = await axios.get(`${vuesionRepo}/tags`);
-    const latestVersion: string = tagsResponse.data[0].name;
+    const tagsResponse: AxiosResponse = await axios.get<any>(`${vuesionRepo}/tags`);
+    const latestVersion: string = next ? 'next' : tagsResponse.data[0].name;
     const currentVersion: string = VuesionConfig.currentVersion;
 
     if (latestVersion === currentVersion) {
@@ -89,16 +114,18 @@ export async function run() {
 
     Result(`Update from version: ${currentVersion} to version: ${latestVersion}.`);
 
-    const diffResponse: AxiosResponse<any> = await axios.get(
-      `${vuesionRepo}/compare/${currentVersion}...${latestVersion}`,
-    );
+    const diffUrl = next
+      ? `${vuesionRepo}/compare/master...next`
+      : `${vuesionRepo}/compare/${currentVersion}...${latestVersion}`;
 
-    handleFiles(diffResponse.data.files);
+    const diffResponse: AxiosResponse = await axios.get<any>(diffUrl);
+
+    handleFiles(diffResponse.data.files, next ? 'next' : 'master');
 
     setTimeout(() => {
       VuesionConfig.load();
       VuesionConfig.updateCurrentVersion(latestVersion);
-    }, 200);
+    }, 1000);
   } catch (e) {
     logErrorBold(e);
   }
